requirements.txt
aiohttp
cairosvg
DateTime
emoji
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gsearch
gtts
hachoir
heroku3
html-telegraph-poster
httpx[http2]
humanize
IMDbPY
instagrapi
justwatch
lottie
lyricsgenius
moviepy
nekos.py
opencv-python-headless
Pillow
psycopg2
pyfiglet
PyGithub
pymediainfo
pySmartDL
python-barcode
pytz
qrcode
regex
requests
search-engine-parser
selenium
sqlalchemy==1.3.23
telegraph
tgcrypto
validators
vcam
WaifuPicsPython

git+https://github.com/jaisingh007/py-googletrans
git+https://github.com/jaisingh007/Telethon
git+https://github.com/jaisingh007/Wikipedia
git+https://github.com/jaisingh007/yt-dp-fork
git+https://github.com/jaisingh007/yt-search-python
